start user with cocktails and menu

+++ ShakeitServer/Models/CocktailIngredients.cs
@@ -10,6 +10,6 @@ namespace ShakeitServer.Models
         public int Id { get; set; }
         public int CocktailId { get; set; }
         public int IngredientId { get; set; }
-        public double Pour { get; set; }
+        public decimal Pour { get; set; }
     }
 }
